Refactored ShiftState and added URI exception
