+++ app/src/main/kotlin/com/simplemobiletools/keyboard/helpers/Constants.kt
@@ -1,82 +1,5 @@
 package com.simplemobiletools.keyboard.helpers
 
-import android.content.Context
-import android.text.InputType
-import com.simplemobiletools.keyboard.extensions.config
-import com.simplemobiletools.keyboard.helpers.MyKeyboard.Companion.KEYCODE_SPACE
-
-enum class ShiftState {
-    OFF,
-    ON_ONE_CHAR,
-    ON_PERMANENT;
-
-    companion object {
-        private const val MIN_TEXT_LENGTH = 2
-        private val inputTypeExceptions = listOf(
-            InputType.TYPE_TEXT_VARIATION_PASSWORD,
-            InputType.TYPE_TEXT_VARIATION_VISIBLE_PASSWORD,
-            InputType.TYPE_TEXT_VARIATION_WEB_PASSWORD,
-            InputType.TYPE_NUMBER_VARIATION_PASSWORD,
-            InputType.TYPE_TEXT_VARIATION_EMAIL_ADDRESS,
-            InputType.TYPE_TEXT_VARIATION_WEB_EMAIL_ADDRESS
-        )
-        private val endOfSentenceChars: List<Char> = listOf('.', '?', '!')
-
-        fun getDefaultShiftState(context: Context, inputTypeClassVariation: Int): ShiftState {
-            if (isInputTypePasswordOrEmail(inputTypeClassVariation)) {
-                return OFF
-            }
-            return when (context.config.enableSentencesCapitalization) {
-                true -> ON_ONE_CHAR
-                else -> OFF
-            }
-        }
-
-        fun getShiftStateForText(context: Context, inputTypeClassVariation: Int, text: String?): ShiftState {
-            if (isInputTypePasswordOrEmail(inputTypeClassVariation)) {
-                return OFF
-            }
-            return when {
-                shouldCapitalize(context, text) -> {
-                    ON_ONE_CHAR
-                }
-
-                else -> {
-                    OFF
-                }
-            }
-        }
-
-        /**
-         * The function is checking whether there is a need in capitalizing based on the given text
-         * @param context Used for checking current sentences capitalization setting
-         * @param text Last text from the input
-         */
-        fun shouldCapitalize(context: Context, text: String?): Boolean {
-            // check whether it is the first letter in textField
-            if (text.isNullOrEmpty()) {
-                return true
-            }
-
-            if (!context.config.enableSentencesCapitalization) {
-                return false
-            }
-
-            val twoLastSymbols = text.takeLast(2)
-
-            if (twoLastSymbols.length < MIN_TEXT_LENGTH) {
-                return false
-            }
-
-            return endOfSentenceChars.contains(twoLastSymbols.first()) && twoLastSymbols.last().code == KEYCODE_SPACE
-        }
-
-        fun isInputTypePasswordOrEmail(inputTypeVariation: Int): Boolean {
-            return inputTypeVariation in inputTypeExceptions
-        }
-    }
-}
-
 // limit the count of alternative characters that show up at long pressing a key
 const val MAX_KEYS_PER_MINI_ROW = 9
